# lib/api-zod/src/generated/types/createGameBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGameBody {
  /**
   * @minLength 1
   * @maxLength 32
   */
  hostName: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  startingLife?: number;
}