# lib/api-zod/src/generated/types/gameLogEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameLogEntryKind } from "./gameLogEntryKind";

export interface GameLogEntry {
  id: string;
  at: Date;
  kind: GameLogEntryKind;
  message: string;
  actorId?: string;
  targetId?: string;
  amount?: number;
}