# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommanderDamageEntry } from "./commanderDamageEntry";
import type { GameLogEntry } from "./gameLogEntry";
import type { GameStateStatus } from "./gameStateStatus";
import type { Player } from "./player";

export interface GameState {
  id: string;
  code: string;
  status: GameStateStatus;
  startingLife: number;
  turnNumber: number;
  currentTurnPlayerId?: string | null;
  hostId: string;
  players: Player[];
  commanderDamage: CommanderDamageEntry[];
  log: GameLogEntry[];
  createdAt: Date;
}