# lib/api-zod/src/generated/types/gameStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatsBiggestSingleHit } from "./gameStatsBiggestSingleHit";
import type { GameStatsTotalDamageDealtItem } from "./gameStatsTotalDamageDealtItem";

export interface GameStats {
  totalDamageDealt: GameStatsTotalDamageDealtItem[];
  biggestSingleHit?: GameStatsBiggestSingleHit;
  eliminatedCount: number;
  turnsPlayed: number;
}