# lib/api-zod/src/generated/types/joinGameBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JoinGameBody {
  /**
   * @minLength 1
   * @maxLength 32
   */
  name: string;
}