# lib/api-zod/src/generated/types/joinResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameState } from "./gameState";

export interface JoinResult {
  token: string;
  playerId: string;
  game: GameState;
}