# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ManaPool } from "./manaPool";

export interface Player {
  id: string;
  name: string;
  life: number;
  isHost: boolean;
  isConnected: boolean;
  isEliminated: boolean;
  position: number;
  color: string;
  commanderTax: number;
  commanderName: string;
  manaPool: ManaPool;
  poisonCounters: number;
  experienceCounters: number;
}